test(vm-runner): move wycheproof vectors to their own file
The wycheproof block tests the RustCrypto p256 crate's handling of
adversarial signatures, not our own wrapper, so it reads better as a
dedicated sibling module than appended to the main test file. Move
the eight tests and their `run_wycheproof` helper into
`p256_verify_wycheproof.rs` and scope the cryptography-specific
vocabulary (`Wycheproof`, `Shamir`, `FIPS`, `Pubkeys`) to that file
via a `// cspell:words` directive, removing them from the global
cspell.json. Also drop `Microbenchmark`, which was only referenced by
the criterion bench removed earlier in this PR.
