+++ cspell.json
@@ -129,7 +129,6 @@
         "externref",
         "fastforward",
         "finalizable",
-        "FIPS",
         "flamegraph",
         "flatstore",
         "fmtr",
@@ -217,7 +216,6 @@
         "memtries",
         "merkelization",
         "Merkle",
-        "Microbenchmark",
         "merklize",
         "merklized",
         "merklizing",
@@ -293,7 +291,6 @@
         "prefetchers",
         "prefunded",
         "privkey",
-        "Pubkeys",
         "profraw",
         "profraws",
         "prost",
@@ -369,7 +366,6 @@
         "SECP",
         "Seedable",
         "serde",
-        "Shamir",
         "shardnet",
         "Shreyan",
         "Shutdownable",
@@ -468,7 +464,6 @@
         "welford",
         "wiremock",
         "wmem",
-        "Wycheproof",
         "Xarrow",
         "xarrows",
         "xoraddr",
